pyproject.toml
[tool.poetry]
name = "mailsync-daemon-poetry"
version = "0.1.0"
description = "Poetry file for the mailsync daemon project"
authors = ["Matteo Gätzner <m.gaetzner@gmx.de>"]

[tool.poetry.dependencies]
python = "==3.10.3"
systemd-python = "==234"
python-gnupg = "==0.4.8"
daemonize = "==2.5.0"

[tool.poetry.dev-dependencies]
black = "*"
pipenv-to-requirements = "*"
pre-commit = "*"
pylint = "*"

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"
